Add method findTeamById

+++ src/main/java/com/brasileirao/championship/Repositories/TeamsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamsRepository extends JpaRepository<Teams, Integer> {
+public interface TeamsRepository extends JpaRepository<Teams, Long> {
 }
